+++ app/templates/base.html
@@ -1,0 +1,11 @@
+<!doctype html>
+<html>
+    <head>
+      <title>Apply for Visa</title>
+    </head>
+    <body>
+        <div>Application: <a href="/index">Home</a></div>
+        <hr>
+        {% block content %}{% endblock %}
+    </body>
+</html>
