+++ app/templates/form.html
@@ -1,0 +1,27 @@
+{% extends "base.html" %}
+
+{% block content %}
+    <form method="post" enctype="multipart/form-data">
+
+        {{ form.csrf_token }}
+        {{ form.full_name.label }} {{ form.full_name(size=20) }}
+        <br>
+        {{ form.email.label }} {{ form.email(size=20) }}
+        <br>
+        {{ form.passport_number.label }} {{ form.passport_number(size=20) }}
+        <br>
+        {{ form.dob.label }} {{ form.dob(size=20) }}
+        <br>
+        {{ form.face_image.label }} {{ form.face_image(size=20) }}
+        <br>
+        {{ form.submit }}
+        <br>
+        {% if form.full_name.errors %}
+            <ul class="errors">
+            {% for error in form.full_name.errors %}
+                <li>{{ error }}</li>
+            {% endfor %}
+            </ul>
+        {% endif %}
+    </form>
+{% endblock %}
