+++ app/templates/index.html
@@ -1,0 +1,8 @@
+{% extends "base.html" %}
+
+{% block content %}
+  <div class="container">
+    <h1>Apply for Vietnam Visa</h1>
+    <a href="/form" class="apply-button">Apply Now!</a>
+  </div>
+{% endblock %}
