+++ app/templates/success.html
@@ -1,0 +1,12 @@
+{% extends "base.html" %}
+
+{% block content %}
+    <h1>Success!</h1>
+    <div id="unique-id">
+        <p>We've sent you an email! You can access the status of your Visa with your email & Passport Number</p>
+        <p>Email:</p>
+        <strong>{{ email }}</strong>
+        <p>Passport Number:</p>
+        <strong>{{ passport_number }}</strong>
+    </div>
+{% endblock %}
