--- templates/form.html
@@ -1,36 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Visa Application</title>
-
-</head>
-<body>
-
-<form method="post" enctype="multipart/form-data">
-
-    {{ form.csrf_token }}
-    {{ form.full_name.label }} {{ form.full_name(size=20) }}
-    <br>
-    {{ form.email.label }} {{ form.email(size=20) }}
-    <br>
-    {{ form.passport_number.label }} {{ form.passport_number(size=20) }}
-    <br>
-    {{ form.dob.label }} {{ form.dob(size=20) }}
-    <br>
-    {{ form.face_image.label }} {{ form.face_image(size=20) }}
-    <br>
-    {{ form.submit }}
-    <br>
-    {% if form.full_name.errors %}
-        <ul class="errors">
-        {% for error in form.full_name.errors %}
-            <li>{{ error }}</li>
-        {% endfor %}
-        </ul>
-    {% endif %}
-</form>
-
-</body>
-</html>
