--- templates/index.html
@@ -1,49 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <style>
-    body {
-      font-family: Arial, sans-serif;
-      margin: 0;
-      padding: 0;
-      background-color: #f4f4f4;
-    }
-
-    .container {
-      max-width: 600px;
-      margin: 0 auto;
-      text-align: center;
-      padding: 50px;
-    }
-
-    h1 {
-      color: #333;
-    }
-
-    .apply-button {
-      display: inline-block;
-      padding: 15px 30px;
-      font-size: 18px;
-      background-color: #3498db;
-      color: #fff;
-      text-decoration: none;
-      border-radius: 5px;
-      transition: background-color 0.3s ease;
-    }
-
-    .apply-button:hover {
-      background-color: #2980b9;
-    }
-  </style>
-</head>
-<body>
-
-  <div class="container">
-    <h1>Apply for Vietnam Visa</h1>
-    <a href="/form" class="apply-button">Apply Now!</a>
-  </div>
-
-</body>
-</html>
