--- templates/success.html
@@ -1,32 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Flask Template</title>
-    <style>
-        body {
-            font-family: 'Arial', sans-serif;
-            text-align: center;
-            margin: 20px;
-        }
-        h1 {
-            color: #007BFF;
-        }
-        #unique-id {
-            margin-top: 20px;
-            font-size: 1.2em;
-        }
-    </style>
-</head>
-<body>
-    <h1>Success!</h1>
-    <div id="unique-id">
-        <p>We've sent you an email! You can access the status of your Visa with your email & Passport Number</p>
-        <p>Email:</p>
-        <strong>{{ email }}</strong>
-        <p>Passport Number:</p>
-        <strong>{{ passport_number }}</strong>
-    </div>
-</body>
-</html>
